State.js: allow to call invalidate() during update phase
It's convenient to be able to call State.invalidate() during update phase. E.g. it's convenient to be able to use same  setters as during non-update phase to propagate information between states. The main danger is marking dirty state that is not inside update graph that has a consumer inside update graph, because this will require to add such producer to update graph as well, which is not possible. However, it's perfectly safe to invalidate states that are already invalidated during update phase.